fix: disable interactive passwords in SSH for git fetch
This commit fixes a problem where SSH (used by Git) is able to
interactively prompt for a password that can not be redirected to
`/dev/null`.

An interactive SSH password prompt inside the async worker does three
things:
1. Leave the `git fetch` hanging due to waiting for input
2. Capture part of the input to the async worker, thus creating
   nonsense commands
3. Use some characters from future async jobs as a password and
   attempts to login

Even if this was fixed in `zsh-async`, it is still good to disable the
password prompt to avoid issue numero uno.